+++ syncopy/nwanalysis/connectivity_analysis.py
@@ -129,6 +129,11 @@ def connectivityanalysis(data, method="coh", keeptrials=False, output="abs",
         For example :func:`~scipy.signal.windows.kaiser` has
         the additional parameter 'beta'. For multi-tapering use `tapsmofrq` directly.
 
+    Returns
+    -------
+    out : `~syncopy.CrossSpectralData`
+        The analyis result with dims ['time', 'freq', 'channel_i', channel_j']
+
     Examples
     --------
     Coming soon...
@@ -265,7 +270,7 @@ def connectivityanalysis(data, method="coh", keeptrials=False, output="abs",
                                          nSamples=nSamples,
                                          taper=taper,
                                          taper_opt=taper_opt,
-                                         demean_taper=(method == 'granger'),
+                                         demean_taper=False,
                                          polyremoval=polyremoval,
                                          timeAxis=timeAxis,
                                          foi=foi)