+++ syncopy/shared/const_def.py
@@ -20,6 +20,7 @@ spectralConversions = {"pow": lambda x: (x * np.conj(x)).real.astype(np.float32)
 
 #: available tapers of :func:`~syncopy.freqanalysis` and  :func:`~syncopy.connectivity`
 all_windows = windows.__all__
+all_windows.remove("get_window")  # aux. function
 all_windows.remove("exponential")  # not symmetric
 all_windows.remove("hanning")  # deprecated
 all_windows.remove("dpss")  # activated via `tapsmofrq`